LPD-24677 Move to a better fitting dir
