+++ modules/apps/frontend-data-set/frontend-data-set-views-web/src/main/resources/META-INF/resources/js/components/OrderableTable.tsx
@@ -19,7 +19,7 @@ import {HTML5Backend} from 'react-dnd-html5-backend';
 import {FUZZY_OPTIONS} from '../utils/constants';
 import Search from './Search';
 
-import '../../css/OrderableTable.scss';
+import '../../css/components/OrderableTable.scss';
 
 const ROW_DRAGGABLE = 'rowDraggable';
