+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/fds_view/visualization_modes/modes/Cards.d.ts
